+++ examples/sprite.rs
@@ -1,7 +1,7 @@
 //! Demonstrates how to use the bevy_sprite picking backend. This backend simply tests the bounds of
 //! a sprite.
 
-use bevy::prelude::*;
+use bevy::{prelude::*, sprite::Anchor};
 use bevy_mod_picking::prelude::*;
 
 fn main() {
@@ -15,20 +15,77 @@ fn main() {
         .run();
 }
 
-fn move_sprite(time: Res<Time>, mut sprite: Query<&mut Transform, With<Sprite>>) {
-    let mut transform = sprite.single_mut();
-    let new = Vec2 {
-        x: 200.0 * time.elapsed_seconds().sin(),
-        y: 200.0 * (time.elapsed_seconds() * 2.0).sin(),
-    };
-    transform.translation.x = new.x;
-    transform.translation.y = new.y;
+fn move_sprite(
+    time: Res<Time>,
+    mut sprite: Query<&mut Transform, (Without<Sprite>, With<Children>)>,
+) {
+    let t = time.elapsed_seconds() * 0.1;
+    for mut transform in &mut sprite {
+        let new = Vec2 {
+            x: 50.0 * t.sin(),
+            y: 50.0 * (t * 2.0).sin(),
+        };
+        transform.translation.x = new.x;
+        transform.translation.y = new.y;
+    }
 }
 
+/// Set up a scene that tests all sprite anchor types.
 fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
     commands.spawn(Camera2dBundle::default());
-    commands.spawn(SpriteBundle {
-        texture: asset_server.load("images/boovy.png"),
-        ..default()
-    });
+
+    let len = 128.0;
+    let sprite_size = Some(Vec2::splat(len / 2.0));
+
+    commands
+        .spawn(SpatialBundle::default())
+        .with_children(|commands| {
+            for (anchor_index, anchor) in [
+                Anchor::TopLeft,
+                Anchor::TopCenter,
+                Anchor::TopRight,
+                Anchor::CenterLeft,
+                Anchor::Center,
+                Anchor::CenterRight,
+                Anchor::BottomLeft,
+                Anchor::BottomCenter,
+                Anchor::BottomRight,
+                Anchor::Custom(Vec2::new(0.5, 0.5)),
+            ]
+            .iter()
+            .enumerate()
+            {
+                // spawn black square behind sprite to show anchor point
+                commands.spawn(SpriteBundle {
+                    sprite: Sprite {
+                        custom_size: sprite_size,
+                        color: Color::BLACK,
+                        ..default()
+                    },
+                    transform: Transform::from_xyz(
+                        (anchor_index % 3) as f32 * len - len,
+                        (anchor_index / 3) as f32 * len - len,
+                        -1.0,
+                    ),
+                    ..default()
+                });
+
+                commands.spawn(SpriteBundle {
+                    sprite: Sprite {
+                        custom_size: sprite_size,
+                        color: Color::RED,
+                        anchor: anchor.to_owned(),
+                        ..default()
+                    },
+                    texture: asset_server.load("images/boovy.png"),
+                    // 3x3 grid of anchor examples by changing transform
+                    transform: Transform::from_xyz(
+                        (anchor_index % 3) as f32 * len - len,
+                        (anchor_index / 3) as f32 * len - len,
+                        0.0,
+                    ),
+                    ..default()
+                });
+            }
+        });
 }